+++ lib/assignment_2/navigation.dart
@@ -1,3 +1,4 @@
+import 'package:assignments/assignment_2/details_page.dart';
 import 'package:flutter/material.dart';
 
 class NavigationHomePage extends StatefulWidget {
@@ -39,33 +40,6 @@ class _NavigationHomePageState extends State<NavigationHomePage> {
                 }
               },
               child: const Text('Go to Details'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class DetailsPage extends StatelessWidget {
-  final String data;
-  const DetailsPage({super.key, required this.data});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Details')),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Text('Recieved: $data', style: const TextStyle(fontSize: 18)),
-            const SizedBox(height: 20),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.pop(context, 'Thanks from Details!');
-              },
-              child: const Text('Send data back'),
             ),
           ],
         ),
